pyproject.toml
[project]
name = "image-search"
version = "0.1.0"
description = "Image similarity search engine using CLIP + FAISS"
readme = "README.md"
requires-python = "==3.11.*"

dependencies = [
    "numpy>=1.26,<2.0",
    "pandas>=2.1,<3.0",
    "pillow>=10,<13",
    "torch==2.5.1",
    "torchvision==0.20.1",
    "transformers==4.40.2",
    "faiss-cpu>=1.7.4,<2.0.0",
    "fastapi>=0.110,<1.0",
    "uvicorn>=0.27,<1.0",
    "streamlit>=1.30,<2.0",
    "kaggle>=1.6,<3.0",
    "ipykernel>=7.2.0",
    "matplotlib>=3.10.8",
    "aiohttp>=3.13.5",
]

[tool.uv]
prerelease = "disallow"
